added grace period of 50 seconds before rotting
Increased the timeframe check to 50 so that meat that is harvested decays after 50 seconds. When cooled, the timeframe resets to 40, leaving a 10 second window

This leaves more time to harvest meat and put it in a crate.

Will have to change the amount of condition remove if meat HEALTH is to be changed to the amount of seconds it takes to eat the meal (eating should always be -1 per second)