Fixed upgrade duplication, reworked detaching
- Duplication should no longer be possible. Seems to be solved by checking if the container is attached.

- Detaching the Kitchen Counter and Kitchen Pantry is now done by holding a wrench when interacting with the item to show a button. Pressing the button will block inventory access and slowly disassemble it (10s)
This change was done as regular detaching would execute the status effects to remove installed fabricators from the characters position, making it possible to execute the effect out of range to remove a fabricator, thus having it stuck in place.
This new approach executes the StatusEffects from the item's position, removing everything reliably.